pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "transcribr"
version = "0.1"
authors = [
  { name="Jonas Kneifl", email="jonas.kneifl@itm.uni-stuttgart.com" },
]
description = "Effortless Audio / Video Transcription and Subtitle Generation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ffmpeg-python",
    "openai-whisper",
]
keywords = [
    "transcription",
    "subtitle",
    "audio",
    "video",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]

[tool.setuptools]


[project.urls]
Homepage = "https://github.com/jkneifl/transcribr"
Issues = "https://github.com/jkneifl/transcribr/issues"
